feat: [PIDM-265] Refactor main package
