+++ src/main/java/it/gov/pagopa/bizevents/sync/nodo/BizEventsSyncNodoApplication.java
@@ -1,0 +1,14 @@
+package it.gov.pagopa.bizevents.sync.nodo;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@SpringBootApplication
+@EnableScheduling
+public class BizEventsSyncNodoApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(BizEventsSyncNodoApplication.class, args);
+  }
+}
