+++ src/main/java/it/gov/pagopa/bizevents/sync/nodo/config/ResponseValidator.java
@@ -1,6 +1,6 @@
-package it.gov.pagopa.microservice.config;
+package it.gov.pagopa.bizevents.sync.nodo.config;
 
-import it.gov.pagopa.microservice.exception.AppException;
+import it.gov.pagopa.bizevents.sync.nodo.exception.AppException;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.Validator;
 import java.util.Set;
