+++ src/main/java/it/gov/pagopa/bizevents/sync/nodo/config/WebMvcConfiguration.java
@@ -1,7 +1,7 @@
-package it.gov.pagopa.microservice.config;
+package it.gov.pagopa.bizevents.sync.nodo.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import it.gov.pagopa.microservice.model.AppCorsConfiguration;
+import it.gov.pagopa.bizevents.sync.nodo.model.AppCorsConfiguration;
 import lombok.SneakyThrows;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
