+++ src/main/java/it/gov/pagopa/bizevents/sync/nodo/exception/AppError.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.microservice.exception;
+package it.gov.pagopa.bizevents.sync.nodo.exception;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
