+++ src/main/java/it/gov/pagopa/bizevents/sync/nodo/model/AppCorsConfiguration.java
@@ -1,6 +1,7 @@
-package it.gov.pagopa.microservice.model;
+package it.gov.pagopa.bizevents.sync.nodo.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -13,10 +14,10 @@ import lombok.ToString;
 @NoArgsConstructor
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @ToString
+@JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class AppInfo {
+public class AppCorsConfiguration {
 
-  private String name;
-  private String version;
-  private String environment;
+  private String[] origins;
+  private String[] methods;
 }
