+++ src/main/java/it/gov/pagopa/bizevents/sync/nodo/model/ProblemJson.java
@@ -1,8 +1,9 @@
-package it.gov.pagopa.microservice.model;
+package it.gov.pagopa.bizevents.sync.nodo.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
+import it.gov.pagopa.bizevents.sync.nodo.exception.ErrorHandler;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
 import lombok.AllArgsConstructor;
@@ -14,7 +15,7 @@ import lombok.ToString;
 /**
  * Object returned as response in case of an error.
  *
- * <p>See {@link it.gov.pagopa.microservice.exception.ErrorHandler}
+ * <p>See {@link ErrorHandler}
  */
 @Data
 @Builder(toBuilder = true)
