+++ src/main/java/it/gov/pagopa/bizevents/sync/nodo/util/CommonUtility.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.microservice.util;
+package it.gov.pagopa.bizevents.sync.nodo.util;
 
 import java.util.Calendar;
 import java.util.List;
