--- src/main/java/it/gov/pagopa/microservice/Application.java
@@ -1,12 +1,0 @@
-package it.gov.pagopa.microservice; // TODO: refactor the package
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Application {
-
-  public static void main(String[] args) {
-    SpringApplication.run(Application.class, args);
-  }
-}
