+++ src/test/java/it/gov/pagopa/bizevents/sync/nodo/BizEventsSyncNodoApplicationTest.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.microservice;
+package it.gov.pagopa.bizevents.sync.nodo;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class ApplicationTest {
+class BizEventsSyncNodoApplicationTest {
 
   @Test
   void contextLoads() {
