+++ src/test/java/it/gov/pagopa/bizevents/sync/nodo/OpenApiGenerationTest.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.microservice;
+package it.gov.pagopa.bizevents.sync.nodo;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -16,7 +16,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
-@SpringBootTest(classes = Application.class)
+@SpringBootTest(classes = BizEventsSyncNodoApplication.class)
 @AutoConfigureMockMvc
 class OpenApiGenerationTest {
